+++ CHANGELOG.md
@@ -9,11 +9,17 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ## [Unreleased]
 
 ### Added
+
+- `proIndexer`: Template demonstrating a reactor app that maintais a summarised form of an Aggregate's as `RollingState` in a separated `Views` Container [#132](https://github.com/jet/dotnet-templates/pull/132)
+
 ### Changed
 
-- Target `Equinox` v `4.0.0-rc.13`, `Propulsion` v `3.0.0-rc.8.10`, `FsCodec` v `3.0.0-rc.11.1`
+- Target `Equinox` v `4.0.0-rc.13`, `Propulsion` v `3.0.0-rc.8.10`, `FsCodec` v `3.0.0-rc.11.1` [#131](https://github.com/jet/dotnet-templates/pull/131)
 
 ### Removed 
+ 
+- `proCosmosReactor`: Extended/split to become `proIndexer` [#132](https://github.com/jet/dotnet-templates/pull/132)
+
 ### Fixed
 
 <a name="6.3.1"></a>